ai-project-estimator/backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
motor>=3.4.0
scikit-learn>=1.5.0
pandas>=2.2.2
joblib>=1.4.2
requests>=2.32.3
python-dotenv>=1.0.1
radon>=6.0.1
GitPython>=3.1.43
openai>=1.35.3
